requirements.txt
elasticsearch>=7.8.0
Flask>=1.1.2
matplotlib>=3.3.0
netCDF4>=1.5.3
numpy>=1.19.0
pandas>=1.0.5
plotly>=4.9.0
scikit-learn>=0.23.1
scipy>=1.5.1
xarray>=0.16.0
nbformat>=5.0.7
requests>=2.24.0
ipywidgets>=7.6.3
gsw
statsmodels
sklearn
erddap
stockstats
